Rename _serialize module to serialize
